+++ package.json
@@ -4,11 +4,15 @@
 	"description": "![pwa](https://user-images.githubusercontent.com/3104648/28351989-7f68389e-6c4b-11e7-9bf2-e9fcd4977e7a.png)",
 	"main": "index.ts",
 	"scripts": {
-		"build": "tsc",
 		"start": "node --experimental-specifier-resolution=node --loader ts-node/esm ./index.ts",
 		"dev": "npm run dev-server & npm run dev-client",
-		"dev-server": "nodemon -e '*' --exec \"node --require ts-node/register --inspect index.ts\"",
-		"dev-client": "tsc --watch --project ./public/scripts/tsconfig.json"
+		"dev-server": "ts-node --esm index.ts",
+		"dev-serverrr": "nodemon -e '*' --exec \"node --require ts-node/register --inspect index.ts\"",
+		"dev-client": "",
+		"prebuild": "rimraf ./public",
+		"build": "npm-run-all build:*",
+		"build:css": "node build-scripts/build-styles.js",
+		"build:js": "node build-scripts/build-scripts.js"
 	},
 	"nodemonConfig": {
 		"ext": "ts,js,hbs"
@@ -26,9 +30,16 @@
 		"@types/express": "^4.17.17",
 		"@types/node": "^18.15.3",
 		"gulp": "^4.0.2",
+		"gulp-autoprefixer": "^8.0.0",
+		"gulp-clean-css": "^4.3.0",
+		"gulp-concat": "^2.6.1",
 		"gulp-typescript": "^6.0.0-alpha.1",
+		"gulp-uglify": "^3.0.2",
 		"nodemon": "^2.0.21",
+		"npm-run-all": "^4.1.5",
+		"rimraf": "^4.4.1",
 		"ts-node": "^10.9.1",
 		"typescript": "^5.0.2"
-	}
+	},
+	"type": "module"
 }
