+++ routes/routes.ts
@@ -1,8 +1,8 @@
-import filters from '../routes/filters';
-import pokemonDetails from '../routes/pokemonDetails';
-import pokemonList from '../routes/pokemonList';
-import splash from '../routes/splash';
-import error from '../routes/error';
+import filters from '../routes/filters.js';
+import pokemonDetails from '../routes/pokemonDetails.js';
+import pokemonList from '../routes/pokemonList.js';
+import splash from '../routes/splash.js';
+import error from '../routes/error.js';
 
 const routes = [
 	{ path: '/filters', view: filters, viewName: 'FilterView' },
