--- src/scripts/utils/delay.ts
@@ -1,5 +1,0 @@
-const delay = (n: number) => {
-	return new Promise((resolve) => setTimeout(resolve, n));
-};
-
-export { delay };
