+++ types/index.ts
@@ -1,7 +1,7 @@
-export {Url} from './Url';
-export {Pokemon} from './Pokemon';
-export {Species} from './Species';
-export {PokemonTypes} from './PokemonTypes';
-export {FullPokemonDetails} from './FullPokemonDetails';
+export {Url} from './Url.js';
+export {Pokemon} from './Pokemon.js';
+export {Species} from './Species.js';
+export {PokemonTypes} from './PokemonTypes.js';
+export {FullPokemonDetails} from './FullPokemonDetails.js';
 
 // export { Url, Pokemon, Species, PokemonTypes, FullPokemonDetails };
